Refactor orchestrator workflows into focused modules
